fix: name of function from chunk to splitEvery fix: edge case functionality to comply with batch (Highland)
